refactor(global): :recycle: app.ts code change to class style
app.ts code change to class style

Ref: #138
